requirements.txt
gunicorn==0.13.4
lovely.jsonrpc==0.2.1
mongoengine==0.5.2
pymongo==2.1.1
tornado==2.1.1
wsgiref==0.1.2
